Initial framework for unit tests of action manager.
A test_action derives from falco::app::action, and additionally saves
a result to return from the action's run() function.

All test actions use the same run_preserve_order function, which saves
the order in which the actions were run, and a deinit function that
does nothing.

The set of tests run various sets of actions in known orders and
compare the expected run order to the actual run order.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>